--- src/main/java/wooteco/subway/dto/LineUpdateRequest.java
@@ -1,19 +1,0 @@
-package wooteco.subway.dto;
-
-public class LineUpdateRequest {
-    private final String name;
-    private final String color;
-
-    public LineUpdateRequest(String name, String color) {
-        this.name = name;
-        this.color = color;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getColor() {
-        return color;
-    }
-}
