+++ src/main/java/wooteco/subway/dto/line/LineSaveRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dto;
+package wooteco.subway.dto.line;
 
 import wooteco.subway.domain.Line;
 
@@ -9,14 +9,21 @@ public class LineSaveRequest {
     private final Long upStationId;
     private final Long downStationId;
     private final int distance;
+    private final int extraFare;
 
     public LineSaveRequest(final String name, final String color, final Long upStationId, final Long downStationId,
-                           final int distance) {
+                           final int distance, final int extraFare) {
         this.name = name;
         this.color = color;
         this.upStationId = upStationId;
         this.downStationId = downStationId;
         this.distance = distance;
+        this.extraFare = extraFare;
+    }
+
+    public LineSaveRequest(final String name, final String color, final Long upStationId, final Long downStationId,
+                           final int distance) {
+        this(name, color, upStationId, downStationId, distance, 0);
     }
 
     public Line toLine() {
@@ -41,5 +48,9 @@ public class LineSaveRequest {
 
     public int getDistance() {
         return distance;
+    }
+
+    public int getExtraFare() {
+        return extraFare;
     }
 }
